+++ lib/persian_number_utility.dart
@@ -2,3 +2,4 @@ library persian_number_utility;
 
 export 'src/methods.dart';
 export 'src/extensions.dart';
+export 'src/enums.dart';